requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0
soundfile>=0.12.0

# OpenVoice dependencies
openvoice-cli>=0.1.0
TTS>=0.20.0

# Speech recognition and processing
speechrecognition>=3.10.0
pyaudio>=0.2.11
pydub>=0.25.1
webrtcvad>=2.0.10

# NLP and conversation
openai>=1.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Audio processing
pyttsx3>=2.90
gTTS>=2.3.0
playsound>=1.3.0

# Web framework for UI
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
colorama>=0.4.6
rich>=13.0.0

# Emotion detection
fer>=22.12.0
opencv-python>=4.8.0

# Data handling
pandas>=2.0.0
python-dateutil>=2.8.2
pyyaml>=6.0

